% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinysnap-package.R
\docType{package}
\name{shinysnap-package}
\alias{shinysnap}
\alias{shinysnap-package}
\title{shinysnap: Save and Restore the State of 'shiny' Applications}
\description{
What the package does (one paragraph, 2+ sentences).
}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]
}

}
\keyword{internal}
