% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinysnap-package.R
\docType{package}
\name{shinysnap-package}
\alias{shinysnap}
\alias{shinysnap-package}
\title{shinysnap: Save and Restore the State of 'shiny' Applications}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Save the state of applications built with 'shiny', the web application framework by Chang et al. (2026) \doi{10.32614/CRAN.package.shiny}. Users can share their work and continue it in another session. Input values and selected values held by the server are saved in JSON (JavaScript Object Notation) files that can be read and edited by hand. Saved state can be restored without reloading the page or setting up bookmarking. Restoration waits for inputs that appear as the page changes and reports which values were restored, missing, or could not be applied.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/shinysnap/}
  \item \url{https://github.com/nanxstats/shinysnap}
  \item Report bugs at \url{https://github.com/nanxstats/shinysnap/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]
}

}
\keyword{internal}
